Allow pool events to be missing a transaction ID
A transaction ID is still required to trigger announcement, so the initial event
back to the submitter needs to reflect the data passed into the connector.
However, the later events triggered to all nodes technically do not need the
transaction ID - so do not enforce this as a hard requirement.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>