Allow some name fields to be UUIDs
Currently "names" and UUIDs are mutually exclusive in all cases - but this was overly
broad. New behavior is outlined below.

Some "named" objects already have an ID - therefore it's confusing to allow their name to
also be a UUID. These objects still cannot be named with a UUID:
  - organizations
  - nodes
  - groups
  - subscriptions
  - datatypes
  - token pools

For "name" fields that are simple strings, there is no confusion in setting them to a UUID,
and in fact a UUID may be a good choice in some scenarios. These fields can be a UUID:
  - namespaces
  - topics
  - tags

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>