Make retries transitive
Retrying an operation that has already been retried will cause it to look
up the newest copy of the operation, and retry that one. In this way,
retries will always form a single chain, and attempting to re-run any of
them will always add a new one to the end of the chain.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>