Create firefly subscription per namespace
 - Each namespace now has a unique firefly contract subscription

To support a subscription per namespace, the `fireflyContract` config
section has been replaced with the `contracts` array section under the
`multiparty` config.

Example:
```
    multiparty:
      enabled: true
      org:
        name: org0
        description: org0
        key: 0x123456
      contract:
        - location:
            address: 0x4ae50189462b0e5d52285f59929d037f790771a6
          firstEvent: oldest
```

 * location is a yaml object containing blockchain plugin configuration
 * `firstEvent` is replacing `fromBlock`, supporting values of newest or oldest

This PR also introduces a new component, multiparty manager, that is
responsible for managing the active firefly contract and interfacing
with the blockchain plugin on behalf of that contract.

Signed-off-by: Alex Shorsher <alex.shorsher@kaleido.io>