Both "pinned" transaction types should share a batch processor
"batch_pin" and "contract_invoke_pin" pull from the same pool of nonces
for private messages, and thus need to share a single dispatch thread (per
message type). This ensures that the order in which nonces are assigned is
the order in which nonces are actually used (which is critical for ordering).

"unpinned" private messages can continue to be a separate dispatcher, as
they don't require nonces.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>